WIP fix replay / state storage issues
