+++ backend/gisst-server/src/templates/instance_all_listing.liquid
@@ -56,7 +56,7 @@
                     <td>{{ state.record.state_name}}</td>
                     <td>{{ state.record.state_description}}</td>
                     <td>
-                        <a class="btn btn-primary" href="/play/{{ instance.info.instance_id }}" target="_blank">
+                        <a class="btn btn-primary" href="/play/{{ instance.info.instance_id }}?state={{ state.record.state_id }}" target="_blank">
                             <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-play" viewBox="0 0 16 16">
                                 <path d="M10.804 8 5 4.633v6.734L10.804 8zm.792-.696a.802.802 0 0 1 0 1.392l-6.363 3.692C4.713 12.69 4 12.345 4 11.692V4.308c0-.653.713-.998 1.233-.696l6.363 3.692z"/>
                             </svg>
@@ -80,7 +80,7 @@
                     <td>{{ replay.record.replay_name}}</td>
                     <td>{{ replay.record.replay_description}}</td>
                     <td>
-                        <a class="btn btn-primary" href="/play/{{ instance.info.instance_id }}" target="_blank">
+                        <a class="btn btn-primary" href="/play/{{ instance.info.instance_id }}?replay={{ replay.record.replay_id }}" target="_blank">
                             <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-play" viewBox="0 0 16 16">
                                 <path d="M10.804 8 5 4.633v6.734L10.804 8zm.792-.696a.802.802 0 0 1 0 1.392l-6.363 3.692C4.713 12.69 4 12.345 4 11.692V4.308c0-.653.713-.998 1.233-.696l6.363 3.692z"/>
                             </svg>
